+++ src/modules/sbstudio/plugin/props/__init__.py
@@ -1,5 +1,11 @@
 from .color import ColorProperty
 from .frame_range import FrameRangeProperty
 from .formation import FormationProperty
+from .light_effects import LightEffectSelectionProperty
 
-__all__ = ("ColorProperty", "FormationProperty", "FrameRangeProperty")
+__all__ = (
+    "ColorProperty",
+    "FormationProperty",
+    "FrameRangeProperty",
+    "LightEffectSelectionProperty",
+)